requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.2,<3.0
pytest>=8.0,<9.0
psycopg[binary]>=3.1,<4.0
pgvector
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-sqlalchemy
sentry-sdk[fastapi]
ecs-logging
ddtrace
newrelic
